2021-11-06 YJ
출석체크 기능 작업,
회원등급별 이미지 분류,
포인트현황 가져오기 외 기능 수정

+++ src/main/java/com/team4/joopging/member/memberVO/MemberVO.java
@@ -31,7 +31,7 @@ public class MemberVO {
     private String memberZipcode;
     private String memberAddress;
     private int memberPloggingCnt;
-    private int memberPoint;
+    private Long memberPoint;
     private int memberSavedPoint;
     private int memberAttend;
     private int memberAttendStatus;
